# public/api-docs.yaml
openapi: 3.0.3
info:
  title: Assessment API - Video Platform
  description: A comprehensive video sharing platform API with authentication, video management, social features, and content discovery
  version: 1.0.0
  contact:
    name: API Support
    email: support@example.com
servers:
  - url: http://localhost:8000/api
    description: Development server
security:
  - JWT: []
components:
  securitySchemes:
    JWT:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token for authentication
...
paths:
...
  /videos:
    get:
      tags:
        - Videos
      summary: Get public videos
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
            maximum: 50
      responses:
        '200':
          description: Videos retrieved successfully
  /videos/upload:
    post:
      tags:
        - Videos
      summary: Upload a video
      security:
        - JWT: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                video:
                  type: string
                  format: binary
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Video uploaded successfully
  /videos/{id}:
    get:
      tags:
        - Videos
      summary: Get video by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Video retrieved successfully
    put:
      tags:
        - Videos
      summary: Update video
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
                isPublic:
                  type: boolean
      responses:
        '200':
          description: Video updated successfully
    delete:
      tags:
        - Videos
      summary: Delete video
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Video deleted successfully
  /videos/search:
    get:
      tags:
        - Videos
      summary: Search videos
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Search results
  /videos/trending:
    get:
      tags:
        - Videos
      summary: Get trending videos
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Trending videos retrieved
  /videos/user/{userId}:
    get:
      tags:
        - Videos
      summary: Get user's videos
      parameters:
        - name: userId
          in: path
          required: true
          schema:
            type: integer
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: User videos retrieved

  # Video Interactions
  /videos/{id}/like:
    post:
      tags:
        - Video Interactions
      summary: Like a video
      security:
        - JWT: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: Video liked successfully
...
  /videos/{id}/view:
    post:
      tags:
        - Video Interactions
      summary: Record video view
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: View recorded

  # Feed Endpoints
...
  /search/videos:
    get:
      tags:
        - Search
      summary: Search videos with filters
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
            minLength: 2
        - name: duration
          in: query
          schema:
            type: string
        - name: uploadDate
          in: query
          schema:
            type: string
        - name: creator
          in: query
          schema:
            type: string
        - name: sortBy
          in: query
          schema:
            type: string
            default: relevance
        - name: page
          in: query
          schema:
            type: integer
            default: 1
      responses:
        '200':
          description: Search results
...